Notification: If the D-Bus method fails, use QSystemTrayIcon as fallback
To do this, let showNotification() check for D-Bus errors, and return a
boolean whether the notification was successful. If that is not the case, or the
program has been compileds without D-Bus support, use the old
QSystemTrayIcon::showMessage() method. If that fails, give up and complain about
it on the debug log.